refactor: LadderGame 기능 분리
